# docs/AI/BASELINE/OPENAPI_SPEC.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AskZack Local Assistant API",
    "version": "0.2.1",
    "description": "Local developer copilot API for context, workspace, repo, project, and git operations. Author field removed from commit endpoints."
  },
  "servers": [
    {
      "url": "https://askzack-local.ngrok.app"
    }
  ],
  "paths": {
...
    "/api/context/active": {
      "put": {
        "operationId": "updateActiveContext",
        "summary": "Update active runtime context",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string",
                    "description": "Markdown content for runtime context"
                  }
                },
                "required": ["content"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Context updated successfully" }
        }
      }
    },
...
  }
}